feat(compiler): content-addressed asset cache + XXH3-128, drop .osbv version header
Switch ContentHasher/AssetStore/AnimationDetector from SHA-256 to
XXH3-128 (System.IO.Hashing) -- faster and wider than the 64-bit
truncated SHA-256 it replaced, no adversary in this pipeline so
non-cryptographic is fine.

AssetStore's hexNaming layout now names files by content hash
(s/{hash}.png, a/{hash}/f{n}.png) instead of a per-run counter, and
SavePng skips re-encoding when the hash-derived path already exists on
disk -- the asset dir becomes a real cross-run cache. Animation folder
hash covers the whole ordered frame sequence, not a single frame, so
two different sequences sharing one frame's content can't collide.
Verified end to end: two compiles of the same .osbv into the same
asset dir produce byte-identical .osb output, with the second run
untouching every reused file (confirmed via mtime) and finishing
faster (25.1s->19.1s sprite-only fixture, 22.7s->10.5s animation-heavy
fixture).

--min-animation-uniqueness is deliberately left out of any future
auto-tuning: unlike tile-size/hash-quant/tile-tolerance/colors (which
change asset bytes, so a different setting legitimately produces a
different hash), it decides which cache namespace (s/ vs a/) identical
content lands in -- tuning it would make the cache path
non-deterministic for the same input. Already measured
zero-sensitivity on every fixture swept this session, so pinning it
costs nothing.

OsbvParser no longer requires the "OsbV: 1" header line (still
accepted and discarded if present) -- .osbv's command grammar already
matched .osb's byte-for-byte, so the header was the only gap keeping a
bare object list from parsing as one. OsbvWriter stops emitting it.

+++ src/OsbMpeg.Compiler/Analysis/TileTimeline.cs
@@ -71,7 +71,7 @@ public sealed class TileRunTracker
 {
     private readonly byte[] _canonicalScratch;
     private readonly bool _canonicalSnapshot;
-    private readonly ulong?[] _currentHash;
+    private readonly UInt128?[] _currentHash;
     private readonly TileGrid _grid;
     private readonly int _hashQuantLevels;
     private readonly byte[]?[] _runSnapshot;
@@ -101,7 +101,7 @@ public sealed class TileRunTracker
         _hashQuantLevels = hashQuantLevels;
         _canonicalSnapshot = canonicalSnapshot;
         _tolerance = tolerance;
-        _currentHash = new ulong?[grid.TileCount];
+        _currentHash = new UInt128?[grid.TileCount];
         _runStartMs = new double[grid.TileCount];
         _runSnapshot = new byte[grid.TileCount][];
         _scratch = new byte[grid.TileSize * grid.TileSize * 3];
